+++ e2e/tests/contracts.rs
@@ -2155,113 +2155,6 @@ async fn blob_contract_deployment() -> Result<()> {
     Ok(())
 }
 
-#[test]
-fn blob_size_must_be_greater_than_zero() {
-    // given
-    let contract = Contract::regular(vec![0x00], Salt::zeroed(), vec![]);
-
-    // when
-    let err = contract
-        .convert_to_loader(0)
-        .expect_err("should have failed because blob size is 0");
-
-    // then
-    assert_eq!(
-        err.to_string(),
-        "blob size must be greater than 0".to_string()
-    );
-}
-
-#[test]
-fn contract_with_no_code_cannot_be_turned_into_a_loader() {
-    // given
-    let contract = Contract::regular(vec![], Salt::zeroed(), vec![]);
-
-    // when
-    let err = contract
-        .convert_to_loader(100)
-        .expect_err("should have failed because there is no code");
-
-    // then
-    assert_eq!(
-        err.to_string(),
-        "must provide at least one blob".to_string()
-    );
-}
-
-#[test]
-fn loader_needs_at_least_one_blob() {
-    // given
-    let no_blobs = vec![];
-
-    // when
-    let err = Contract::loader_for_blobs(no_blobs, Salt::default(), vec![])
-        .expect_err("should have failed because there are no blobs");
-
-    // then
-    assert_eq!(
-        err.to_string(),
-        "must provide at least one blob".to_string()
-    );
-}
-
-#[test]
-fn loader_requires_all_except_the_last_blob_to_be_word_sized() {
-    // given
-    let blobs = [vec![0; 9], vec![0; 8]].map(Blob::new).to_vec();
-
-    // when
-    let err = Contract::loader_for_blobs(blobs, Salt::default(), vec![])
-        .expect_err("should have failed because the first blob is not word-sized");
-
-    // then
-    assert_eq!(
-        err.to_string(),
-        "blob 1/2 has a size of 9 bytes, which is not a multiple of 8".to_string()
-    );
-}
-
-#[test]
-fn last_blob_in_loader_can_be_unaligned() {
-    // given
-    let blobs = [vec![0; 8], vec![0; 9]].map(Blob::new).to_vec();
-
-    // when
-    let result = Contract::loader_for_blobs(blobs, Salt::default(), vec![]);
-
-    // then
-    let _ = result.unwrap();
-}
-
-#[test]
-fn can_load_regular_contract() -> Result<()> {
-    // given
-    let contract_binary = "sway/contracts/contract_test/out/release/contract_test.bin";
-
-    // when
-    let contract = Contract::load_from(contract_binary, LoadConfiguration::default())?;
-
-    // then
-    assert_eq!(contract.code(), std::fs::read(contract_binary)?);
-
-    Ok(())
-}
-
-#[test]
-fn can_manually_create_regular_contract() -> Result<()> {
-    // given
-    let contract_binary = "sway/contracts/contract_test/out/release/contract_test.bin";
-    let binary = std::fs::read(contract_binary)?;
-
-    // when
-    let contract = Contract::regular(binary.clone(), Salt::zeroed(), vec![]);
-
-    // then
-    assert_eq!(contract.code(), binary);
-
-    Ok(())
-}
-
 #[tokio::test]
 async fn regular_contract_can_be_deployed() -> Result<()> {
     // given
@@ -2295,78 +2188,6 @@ async fn regular_contract_can_be_deployed() -> Result<()> {
     Ok(())
 }
 
-macro_rules! getters_work {
-    ($contract: ident, $contract_id: expr, $state_root: expr, $code_root: expr, $salt: expr, $code: expr) => {
-        assert_eq!($contract.contract_id(), $contract_id);
-        assert_eq!($contract.state_root(), $state_root);
-        assert_eq!($contract.code_root(), $code_root);
-        assert_eq!($contract.salt(), $salt);
-        assert_eq!($contract.code(), $code);
-    };
-}
-
-#[test]
-fn regular_contract_has_expected_getters() -> Result<()> {
-    let contract_binary = "sway/contracts/contract_test/out/release/contract_test.bin";
-    let contract = Contract::load_from(contract_binary, LoadConfiguration::default())?;
-    let expected_binary = std::fs::read(contract_binary)?;
-    let expected_contract_id =
-        "d62e0d1326dd2aca749657c59c3288c0bdd0fe664482f72d3a85a9e289ff92cb".parse()?;
-    let expected_state_root =
-        "8aaf36bbaf5ef509ba6013e11b0f7176de429c9b6b3c1b09cc50b36e3113b901".parse()?;
-    let expected_code_root =
-        "d6721f6198f3765615222390655ce1acfe2ce1ec06002527cfa786b6d705ad0b".parse()?;
-    let expected_salt = Salt::zeroed();
-
-    getters_work!(
-        contract,
-        expected_contract_id,
-        expected_state_root,
-        expected_code_root,
-        expected_salt,
-        expected_binary
-    );
-
-    Ok(())
-}
-
-#[test]
-fn regular_can_be_turned_into_loader_and_back() -> Result<()> {
-    let contract_binary = "sway/contracts/contract_test/out/release/contract_test.bin";
-
-    let contract_original = Contract::load_from(contract_binary, LoadConfiguration::default())?;
-
-    let loader_contract = contract_original.clone().convert_to_loader(1024)?;
-
-    let regular_recreated = loader_contract.clone().revert_to_regular();
-
-    assert_eq!(regular_recreated, contract_original);
-
-    Ok(())
-}
-
-#[test]
-fn unuploaded_loader_contract_has_expected_getters() -> Result<()> {
-    let contract_binary = "sway/contracts/contract_test/out/release/contract_test.bin";
-
-    let original = Contract::load_from(contract_binary, LoadConfiguration::default())?;
-    let loader = original.clone().convert_to_loader(1024)?;
-
-    let loader_asm = loader_contract_asm(&loader.blob_ids()).unwrap();
-    let manual_loader = original.with_code(loader_asm);
-
-    getters_work!(
-        loader,
-        manual_loader.contract_id(),
-        manual_loader.state_root(),
-        manual_loader.code_root(),
-        manual_loader.salt(),
-        manual_loader.code()
-    );
-
-    Ok(())
-}
-
 #[tokio::test]
 async fn unuploaded_loader_can_be_deployed_directly() -> Result<()> {
     setup_program_test!(
@@ -2393,52 +2214,6 @@ async fn unuploaded_loader_can_be_deployed_directly() -> Result<()> {
     Ok(())
 }
 
-#[test]
-fn unuploaded_loader_requires_at_least_one_blob() -> Result<()> {
-    // given
-    let no_blob_ids = vec![];
-
-    // when
-    let loader = Contract::loader_for_blob_ids(no_blob_ids, Salt::default(), vec![])
-        .expect_err("should have failed because there are no blobs");
-
-    // then
-    assert_eq!(
-        loader.to_string(),
-        "must provide at least one blob".to_string()
-    );
-    Ok(())
-}
-
-#[test]
-fn uploaded_loader_has_expected_getters() -> Result<()> {
-    let contract_binary = "sway/contracts/huge_contract/out/release/huge_contract.bin";
-    let original_contract = Contract::load_from(contract_binary, LoadConfiguration::default())?;
-
-    let blob_ids = original_contract
-        .clone()
-        .convert_to_loader(1024)?
-        .blob_ids();
-
-    // we pretend we uploaded the blobs
-    let loader = Contract::loader_for_blob_ids(blob_ids.clone(), Salt::default(), vec![])?;
-
-    let loader_asm = loader_contract_asm(&blob_ids).unwrap();
-    let manual_loader = original_contract.with_code(loader_asm);
-
-    getters_work!(
-        loader,
-        manual_loader.contract_id(),
-        manual_loader.state_root(),
-        manual_loader.code_root(),
-        manual_loader.salt(),
-        manual_loader.code()
-    );
-
-    Ok(())
-}
-
-//
 #[tokio::test]
 async fn unuploaded_loader_can_upload_blobs_separately_then_deploy() -> Result<()> {
     setup_program_test!(
