+++ packages/fuels-programs/src/contract.rs
@@ -56,7 +56,10 @@ fn compute_contract_id_and_state_root(
 mod tests {
     use std::path::Path;
 
-    use fuels_core::types::errors::Error;
+    use fuels_core::types::{
+        errors::{Error, Result},
+        transaction_builders::Blob,
+    };
     use tempfile::tempdir;
 
     use super::*;
@@ -111,5 +114,237 @@ mod tests {
             serde_json::to_string::<Vec<StorageSlot>>(slots).unwrap(),
         )
         .unwrap()
+    }
+
+    #[test]
+    fn blob_size_must_be_greater_than_zero() {
+        // given
+        let contract = Contract::regular(vec![0x00], Salt::zeroed(), vec![]);
+
+        // when
+        let err = contract
+            .convert_to_loader(0)
+            .expect_err("should have failed because blob size is 0");
+
+        // then
+        assert_eq!(
+            err.to_string(),
+            "blob size must be greater than 0".to_string()
+        );
+    }
+
+    #[test]
+    fn contract_with_no_code_cannot_be_turned_into_a_loader() {
+        // given
+        let contract = Contract::regular(vec![], Salt::zeroed(), vec![]);
+
+        // when
+        let err = contract
+            .convert_to_loader(100)
+            .expect_err("should have failed because there is no code");
+
+        // then
+        assert_eq!(
+            err.to_string(),
+            "must provide at least one blob".to_string()
+        );
+    }
+
+    #[test]
+    fn loader_needs_at_least_one_blob() {
+        // given
+        let no_blobs = vec![];
+
+        // when
+        let err = Contract::loader_for_blobs(no_blobs, Salt::default(), vec![])
+            .expect_err("should have failed because there are no blobs");
+
+        // then
+        assert_eq!(
+            err.to_string(),
+            "must provide at least one blob".to_string()
+        );
+    }
+
+    #[test]
+    fn loader_requires_all_except_the_last_blob_to_be_word_sized() {
+        // given
+        let blobs = [vec![0; 9], vec![0; 8]].map(Blob::new).to_vec();
+
+        // when
+        let err = Contract::loader_for_blobs(blobs, Salt::default(), vec![])
+            .expect_err("should have failed because the first blob is not word-sized");
+
+        // then
+        assert_eq!(
+            err.to_string(),
+            "blob 1/2 has a size of 9 bytes, which is not a multiple of 8".to_string()
+        );
+    }
+
+    #[test]
+    fn last_blob_in_loader_can_be_unaligned() {
+        // given
+        let blobs = [vec![0; 8], vec![0; 9]].map(Blob::new).to_vec();
+
+        // when
+        let result = Contract::loader_for_blobs(blobs, Salt::default(), vec![]);
+
+        // then
+        let _ = result.unwrap();
+    }
+
+    #[test]
+    fn can_load_regular_contract() -> Result<()> {
+        // given
+        let tmp_dir = tempfile::tempdir()?;
+        let code_file = tmp_dir.path().join("contract.bin");
+        let code = b"some fake contract code";
+        std::fs::write(&code_file, code)?;
+
+        // when
+        let contract = Contract::load_from(
+            code_file,
+            LoadConfiguration::default()
+                .with_storage_configuration(StorageConfiguration::default().with_autoload(false)),
+        )?;
+
+        // then
+        assert_eq!(contract.code(), code);
+
+        Ok(())
+    }
+
+    #[test]
+    fn can_manually_create_regular_contract() -> Result<()> {
+        // given
+        let binary = b"some fake contract code";
+
+        // when
+        let contract = Contract::regular(binary.to_vec(), Salt::zeroed(), vec![]);
+
+        // then
+        assert_eq!(contract.code(), binary);
+
+        Ok(())
+    }
+
+    macro_rules! getters_work {
+        ($contract: ident, $contract_id: expr, $state_root: expr, $code_root: expr, $salt: expr, $code: expr) => {
+            assert_eq!($contract.contract_id(), $contract_id);
+            assert_eq!($contract.state_root(), $state_root);
+            assert_eq!($contract.code_root(), $code_root);
+            assert_eq!($contract.salt(), $salt);
+            assert_eq!($contract.code(), $code);
+        };
+    }
+
+    #[test]
+    fn regular_contract_has_expected_getters() -> Result<()> {
+        let contract_binary = b"some fake contract code";
+        let storage_slots = vec![StorageSlot::new([2; 32].into(), [1; 32].into())];
+        let contract = Contract::regular(contract_binary.to_vec(), Salt::zeroed(), storage_slots);
+
+        let expected_contract_id =
+            "93c9f1e61efb25458e3c56fdcfee62acb61c0533364eeec7ba61cb2957aa657b".parse()?;
+        let expected_state_root =
+            "852b7b7527124dbcd44302e52453b864dc6f4d9544851c729da666a430b84c97".parse()?;
+        let expected_code_root =
+            "69ca130191e9e469f1580229760b327a0729237f1aff65cf1d076b2dd8360031".parse()?;
+        let expected_salt = Salt::zeroed();
+
+        getters_work!(
+            contract,
+            expected_contract_id,
+            expected_state_root,
+            expected_code_root,
+            expected_salt,
+            contract_binary
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn regular_can_be_turned_into_loader_and_back() -> Result<()> {
+        let contract_binary = b"some fake contract code";
+
+        let contract_original = Contract::regular(contract_binary.to_vec(), Salt::zeroed(), vec![]);
+
+        let loader_contract = contract_original.clone().convert_to_loader(1)?;
+
+        let regular_recreated = loader_contract.clone().revert_to_regular();
+
+        assert_eq!(regular_recreated, contract_original);
+
+        Ok(())
+    }
+
+    #[test]
+    fn unuploaded_loader_contract_has_expected_getters() -> Result<()> {
+        let contract_binary = b"some fake contract code";
+
+        let storage_slots = vec![StorageSlot::new([2; 32].into(), [1; 32].into())];
+        let original = Contract::regular(contract_binary.to_vec(), Salt::zeroed(), storage_slots);
+        let loader = original.clone().convert_to_loader(1024)?;
+
+        let loader_asm = loader_contract_asm(&loader.blob_ids()).unwrap();
+        let manual_loader = original.with_code(loader_asm);
+
+        getters_work!(
+            loader,
+            manual_loader.contract_id(),
+            manual_loader.state_root(),
+            manual_loader.code_root(),
+            manual_loader.salt(),
+            manual_loader.code()
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn unuploaded_loader_requires_at_least_one_blob() -> Result<()> {
+        // given
+        let no_blob_ids = vec![];
+
+        // when
+        let loader = Contract::loader_for_blob_ids(no_blob_ids, Salt::default(), vec![])
+            .expect_err("should have failed because there are no blobs");
+
+        // then
+        assert_eq!(
+            loader.to_string(),
+            "must provide at least one blob".to_string()
+        );
+        Ok(())
+    }
+
+    #[test]
+    fn uploaded_loader_has_expected_getters() -> Result<()> {
+        let contract_binary = b"some fake contract code";
+        let original_contract = Contract::regular(contract_binary.to_vec(), Salt::zeroed(), vec![]);
+
+        let blob_ids = original_contract
+            .clone()
+            .convert_to_loader(1024)?
+            .blob_ids();
+
+        // we pretend we uploaded the blobs
+        let loader = Contract::loader_for_blob_ids(blob_ids.clone(), Salt::default(), vec![])?;
+
+        let loader_asm = loader_contract_asm(&blob_ids).unwrap();
+        let manual_loader = original_contract.with_code(loader_asm);
+
+        getters_work!(
+            loader,
+            manual_loader.contract_id(),
+            manual_loader.state_root(),
+            manual_loader.code_root(),
+            manual_loader.salt(),
+            manual_loader.code()
+        );
+
+        Ok(())
     }
 }
